Changed name of two tests.
Changed the name of two tests to use plural forms. This better represents the two assertions it makes. One might argue its bad style to put two assertions in one test and that the better approach is to use two methods, however I feel that here the approach makes sense.